requirement1.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens 
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
colorama 
coloredlogs==15.0.1
comm 
dataclasses-json==0.6.7
debugpy 
decorator 
dill==0.4.0
distro==1.9.0
docling==2.47.1
docling-core==2.45.0
docling-ibm-models==3.9.0
docling-parse==4.2.3
durationpy==0.10
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup 
executing 
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watson_machine_learning==1.0.368
idna==3.10
imageio==2.37.0
importlib_metadata 
importlib_resources==6.5.2
ipykernel 
ipython 
ipython_pygments_lexers 
jedi 
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client 
jupyter_core 
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.16
latex2mathml==3.78.0
lazy_loader==0.4
lomond==0.3.3
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline 
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nest_asyncio 
networkx==3.5
ninja==1.13.0
#numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging 
pandas==2.1.4
parso 
pathlib==1.0.1
pickleshare 
pillow==11.3.0
platformdirs 
posthog==5.4.0
prompt_toolkit 
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil 
pure_eval 
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments 
pylatexenc==2.10
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil 
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pywin32 
PyYAML==6.0.2
pyzmq 
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semchunk==2.2.2
sentence-transformers==5.1.0
shapely==2.1.1
shellingham==1.5.4
six 
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack_data 
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tornado 
tqdm==4.67.1
traitlets 
transformers==4.55.4
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions 
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth 
websocket-client==1.8.0
websockets==15.0.1
xlsxwriter==3.2.5
yarl==1.20.1
zipp 
zstandard==0.24.0
#opencv-python-headless==4.9.0.80
langchain
langchain-community
langchain-chroma
langchain-ibm
langchain-huggingface
